Move the optional key field set/get to ARCH specific (ex. psa) include/tdi/arch/tna/tna_defs.h - since it is a part of core already.     remove the TDI_MATCH_TYPE_TNA_RANGE include/tdi/arch/tna/tna_info.hpp     remote the string RANGE definition, since it is a part of core already. include/tdi/common/c_frontend/tdi_table_key.h     move the OPTIONAL key field set/get to arch specific(ex. psa). include/tdi/common/tdi_info.hpp     add string for TABLE_KEY_MATCH_TYPE_RANGE = "Range"     add the matchEnumMapAdd with RANGE in the core. include/tdi/common/tdi_table_key.hpp     remove the enum KeyFieldType and map     add range to enum tdi_match_type_core_e     move the KeyFieldValueOptional to arch specific (ex. psa) src/CMakeLists.txt     add subdirectory arch/psa/c_frontend. src/c_frontend/tdi_table_key_c.cpp     move the optional implementation to arch specific. new file:   include/tdi/arch/psa/c_frontend/psa_table_key.h     add tdi_match_type_spa_e with TDI_MATCH_TYPE_PSA_OPTIONAL     add function prototypes for set/get optional new file:   include/tdi/arch/psa/psa_table_key.hpp     add optional KeyFieldValueOptional definition for PSA KeyFieldValue new file:   src/arch/psa/CMakeLists.txt     add tdi_psa(with psa src including c_frontend) lib link with tdi new file:   src/arch/psa/c_frontend/psa_table_key_c.cpp     moved the optional keyFieldValue implementation to this psa specific file.
