+++ src/c_frontend/tdi_table_key_c.cpp
@@ -119,26 +119,6 @@ tdi_status_t tdi_key_field_set_value_lpm_ptr(tdi_table_key_hdl *key_hdl,
   return key->setValue(field_id, keyFieldValue);
 }
 
-/* Optional */
-tdi_status_t tdi_key_field_set_value_optional(tdi_table_key_hdl *key_hdl,
-                                               const tdi_id_t field_id,
-                                               const uint64_t value1,
-                                               const bool is_valid) {
-  auto key = reinterpret_cast<tdi::TableKey *>(key_hdl);
-  tdi::KeyFieldValueOptional<const uint8_t> keyFieldValue(value1, is_valid);
-  return key->setValue(field_id, keyFieldValue);
-}
-
-tdi_status_t tdi_key_field_set_value_optional_ptr(tdi_table_key_hdl *key_hdl,
-                                                   const tdi_id_t field_id,
-                                                   const uint8_t *value1,
-                                                   const bool is_valid,
-                                                   const size_t size) {
-  auto key = reinterpret_cast<tdi::TableKey *>(key_hdl);
-  tdi::KeyFieldValueOptional<const uint8_t> keyFieldValue(value1, is_valid, size);
-  return key->setValue(field_id, keyFieldValue);
-}
-
 /** Get */
 /* Exact */
 tdi_status_t tdi_key_field_get_value(const tdi_table_key_hdl *key_hdl,
@@ -243,27 +223,5 @@ tdi_status_t tdi_key_field_get_value_lpm_ptr(
     const uint16_t *p_length) {
   auto key = reinterpret_cast<const tdi::TableKey *>(key_hdl);
   tdi::KeyFieldValueLPM<unsigned char> keyFieldValue(value1, *p_length, size);
-  return key->getValue(field_id, &keyFieldValue);
-}
-
-/* OPTIONAL */
-tdi_status_t tdi_key_field_get_value_optional(
-    const tdi_table_key_hdl *key_hdl,
-    const tdi_id_t field_id,
-    uint64_t *value,
-    bool *is_valid) {
-  auto key = reinterpret_cast<const tdi::TableKey *>(key_hdl);
-  tdi::KeyFieldValueOptional<long unsigned int> keyFieldValue(*value, *is_valid);
-  return key->getValue(field_id, &keyFieldValue);
-}
-
-tdi_status_t tdi_key_field_get_value_optional_ptr(
-    const tdi_table_key_hdl *key_hdl,
-    const tdi_id_t field_id,
-    const size_t size,
-    uint8_t *value1,
-    bool *is_valid) {
-  auto key = reinterpret_cast<const tdi::TableKey *>(key_hdl);
-  tdi::KeyFieldValueOptional<unsigned char> keyFieldValue(value1, *is_valid, size);
   return key->getValue(field_id, &keyFieldValue);
 }
